+++ rio-backend/src/config/window.rs
@@ -14,10 +14,9 @@ pub enum WindowMode {
     Windowed,
 }
 
-#[derive(Clone, Default, Serialize, Deserialize, Copy, Debug, PartialEq)]
+#[derive(Clone, Serialize, Deserialize, Copy, Debug, PartialEq)]
 pub enum Decorations {
     #[serde(alias = "enabled")]
-    #[default]
     Enabled,
     #[serde(alias = "disabled")]
     Disabled,
@@ -25,6 +24,19 @@ pub enum Decorations {
     Transparent,
     #[serde(alias = "buttonless")]
     Buttonless,
+}
+
+#[allow(clippy::derivable_impls)]
+impl Default for Decorations {
+    fn default() -> Decorations {
+        #[cfg(target_os = "macos")]
+        {
+            Decorations::Transparent
+        }
+
+        #[cfg(not(target_os = "macos"))]
+        Decorations::Enabled
+    }
 }
 
 #[derive(PartialEq, Serialize, Deserialize, Clone, Debug)]